devil/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
ddgs
python-dotenv
redis
requests
beautifulsoup4
urllib3
